fixup! Fix 32-bit Windows issues related to QUIC_ACKM / QUIC_CC
This is a reversal of some of the size_t → unint64_t changes.  They were
made to counter some build failures, but may have been an over-reaction.
Extra attention needs to be payed with what happens when this is added.

+++ include/internal/quic_wire.h
@@ -116,7 +116,7 @@ typedef struct ossl_quic_frame_ack_st {
      * acknowledge at least one packet number.
      */
     OSSL_QUIC_ACK_RANGE        *ack_ranges;
-    uint64_t                    num_ack_ranges;
+    size_t                      num_ack_ranges;
 
     OSSL_TIME                   delay_time;
     uint64_t                    ect0, ect1, ecnce;
